Better terminal module
Through the terminal module you can now use a basic version of the classic C function called printf. This function is FAR from being comparable to the original printf, but it's a good start and very useful for debugging and testing(at least I don't have to convert a number to string every time I want to print it on screen).
It includes: %d(for integers), %c(for characters), %s(for C strings).

Also made some other minor change.