Tape module
The tape module contains functions useful for tape manipulation. The given demo prints every block from the start of the tape to its end.
No tape is given, so you should create your own. To do so just run the following:

  umps2-mkdev -t tape_name.umps file1 file2 ...

tape_name is an arbitrary name, and is the given to the tape once its created. file1, file2 etc. are the list of files you want to include in your tape.
Once your tape is ready you can include it in umps2 from:  machine -> edit configuration