Tests
Created the tests module. Here's where we are going to put all the examples for the phase0 of the project.
Introduces a VM directory in the project structure, where I put a preconfigured virtual machine called 'machine.umps' and the first demo tape.

IMPORTANT: don't move the machine from there nor change the project structure or the configuration might break. The pros of this approach, however, are almost obvious. You just need to open the machine from umps and start the program and it'll do everything automagically.
.